+++ src/lib/AlphaSlider.svelte
@@ -3,6 +3,9 @@
   import Slider from './Slider.svelte'
 
   export let color: Color
+  export let onInput: (color: number) => void = () => {
+    /* noop */
+  }
 </script>
 
 <Slider
@@ -11,4 +14,5 @@
   max={1}
   handleColor="hsla({color.h},{color.s * 100}%,{color.v * 100}%,{color.a})"
   style="alpha"
+  {onInput}
 />